Implemented script functions runto and enablemouse. Implemented special variable actiontext.
git-svn-id: svn://svn.code.sf.net/p/captainsengine/code/trunk@807 dc34a4ac-3e42-4537-918b-2dea8b01751e

+++ Adventure/AppBuilder/AppBuilder/Animator.cpp
@@ -15,7 +15,7 @@ Animator::~Animator(){
 
 }
 
-void Animator::add(Object2D* obj, const std::list<Vec2i>& targetpath, int speedfactor){
+void Animator::add(Object2D* obj, const std::list<Vec2i>& targetpath, float speedfactor){
   std::map<Object2D*,ObjectAnim>::iterator iter = mObjects.find(obj);
   if (iter != mObjects.end()){
     if (!targetpath.empty())
